givbox-main/requirements.txt
asgiref==3.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
django-filter==22.1
django-fsm==2.8.0
djangorestframework~=3.14.0
drf-yasg
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.8
firebase_admin
psycopg2-binary
django-cors-headers

Django~=4.1.3
djangorestframework-simplejwt
django-jazzmin==2.6.0
